@@ -1,14 +1,31 @@
 [build-system]
-requires      = ["setuptools>=61.0.0", "wheel", "setuptools_scm[toml]>=6.2"]
+requires      = ["setuptools>=64.0.0", "wheel", "setuptools_scm>=6.2.0"]
 build-backend = "setuptools.build_meta"
 
-[tool.setuptools_scm]
+[tool.black]
+line-length = 88
+include = '/src/.*\.pyi?$|/tests/.*\.pyi?$|/docs/.*\.pyi?$|noxfile.py'
+
+classifiers = [
+    'Natural Language :: English',
+    'Operating System :: Raspberry Pi',
+    'Programming Language :: Python',
+    'Programming Language :: Python :: 3 :: Only',
+    'Programming Language :: Python :: 3',
+    'Typing :: Typed',
+]
+
+
+[tool.ruff.isort]
+known-first-party = [
+    "astropitography",
+]
 
 [project]
 name = "astropitography"
 description = "Astrophotography + raspberry pi HQ camera = AstroPitography"
 readme = "README.md"
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 keywords = ["astrophotography", "raspberry pi"]
 authors = [
     { name = "Adam Baskerville", email = "adamlukebaskerville@gmail.com" },
